backend/agent/requirements.txt
browser_use==0.1.36
Flask==3.1.0
Flask_Cors==5.0.0
langchain==0.3.14
python-dotenv==1.0.1
